requirements.txt
sentence-transformers
faiss-cpu
gradio
gitpython
numpy
pandas
